requirements.txt
argon2-cffi==20.1.0
base58==2.0.0
cffi==1.14.0
pycparser==2.20
pycryptodome==3.9.7
PyNaCl==1.3.0
six==1.14.0
